backend/requirements.txt
uvicorn
google-cloud-aiplatform>=1.35.0
google-genai
google-generativeai
google-cloud-documentai
google-api-core
fastapi
pydantic
loguru
PyPDF2
pdfminer.six
pdfplumber
pdf2image
Pillow
pytesseract
langchain-core
langchain-google-vertexai
opencv-python
python-multipart
